Fixed CUDA includes and files

+++ kernel/antennaGain.cu
@@ -1,6 +1,3 @@
-#include "cuda_runtime.h"
-#include "device_launch_parameters.h"
-#include "maths_constants.h"
 #include "utils/sinc.cu"
 
 __device__ float AntennaGain(
@@ -11,7 +8,7 @@ __device__ float AntennaGain(
     float az_sinc = sincf(azimuth / beam_width);
     float el_sinc = sincf(elevation / beam_width);
 
-    float gain = pow(az_sinc, 2) * pow(el_sinc, 2);
+    float gain = powf(az_sinc, 2) * powf(el_sinc, 2);
     
     return gain;
 }
